Image capture script (#38)
Wrote my code that starts taking photos after the Raspberry Pi turns on. I originally wanted to also connect it to the Autonomy.py file as that contains the information for the mission status (start/pause/end).

I've also tested the code using one of the Pi's and a camera, and it starts taking photos and stores them to a directory on the pi I was using.

There were other camera options that should be experimented with. A question I had while I was writing this short script were:
"When are missions typically going to take place"
Because that has a slight impact on the different optical settings we can set for the camera in terms of it's color correcting/focus parameters.